backend/diarization/requirements_diarization.txt
annotated-types==0.6.0
anyio==3.7.1
av==10.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coloredlogs==15.0.1
ctranslate2==3.20.0
exceptiongroup==1.1.3
fastapi==0.103.2
faster-whisper==0.9.0
filelock==3.12.4
flatbuffers==23.5.26
fsspec==2023.9.2
h11==0.14.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
mpmath==1.3.0
numpy==1.26.0
onnxruntime==1.16.0
packaging==23.2
protobuf==4.24.4
pyannote.audio==3.0.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.14.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
